refactor: extract shared SavedObjectRenderer component
DynamicScenePage and ComicBookReader both defined near-identical
SavedObjectRenderer helpers that looked up a card type in
CARD_TYPE_REGISTRY, rendered its content, and wrapped it in a
SceneObject. Extract to src/components/scene/SavedObjectRenderer.jsx
and import from both pages.

The shared component accepts the union of props. Editor-only props
(selected, overridePosition, onSelect, onDragStart) are optional;
when onSelect is absent the component falls back to read-only
rendering with onObjectClick as the plain click handler. Default
behavior is read-only, matching the reader.

Closes #80
