+++ src/components/scene/index.js
@@ -3,3 +3,4 @@ export { SceneObject, ObjectPresets } from './SceneObject';
 export { SceneObjectGroup, useGroup } from './SceneObjectGroup';
 export { Panel } from './Panel';
 export { InsertToolbar } from './InsertToolbar';
+export { SavedObjectRenderer } from './SavedObjectRenderer';
